+++ ravyn/routing/controllers/base.py
@@ -145,10 +145,10 @@ class BaseController:
             **Example**
 
             ```python
-            from ravyn import APIView, JSONResponse
+            from ravyn import Controller, JSONResponse
 
 
-            class MyView(APIView):
+            class MyView(Controller):
                 response_class = JSONResponse
             ```
             """
@@ -165,7 +165,7 @@ class BaseController:
             **Example**
 
             ```python
-            from ravyn import APIView
+            from ravyn import Controller
             from ravyn.datastructures import Cookie
 
             response_cookies=[
@@ -177,7 +177,7 @@ class BaseController:
                 )
             ]
 
-            class MyView(APIView):
+            class MyView(Controller):
                 response_cookies = response_cookies
             ```
             """
@@ -194,14 +194,14 @@ class BaseController:
             **Example**
 
             ```python
-            from ravyn import APIView
+            from ravyn import Controller
             from ravyn.datastructures import ResponseHeader
 
             response_headers={
                 "authorize": ResponseHeader(value="granted")
             }
 
-            class MyView(APIView):
+            class MyView(Controller):
                 response_headers = response_headers
             ```
             """