+++ tests/permissions/test_api_views_inheritance.py
@@ -3,7 +3,7 @@ from ravyn.permissions import AllowAny, DenyAll, IsAuthenticated
 from ravyn.testclient import create_client
 
 
-class MyAPIView(Controller):
+class MyController(Controller):
     permissions = [DenyAll]
 
     @get("/home")
@@ -11,7 +11,7 @@ class MyAPIView(Controller):
         return "home"
 
 
-class MySimpleAPIView(MyAPIView):
+class MySimpleController(MyController):
     permissions = [AllowAny]
 
     @get(
@@ -21,27 +21,27 @@ class MySimpleAPIView(MyAPIView):
         return "home simple"
 
 
-class AnotherAPI(MyAPIView):
+class AnotherAPI(MyController):
     permissions = [AllowAny]
 
     @get("/another", permissions=[IsAuthenticated])
     async def get(self) -> str: ...
 
 
-def test_cannot_access_apiview(test_client_factory):
-    with create_client(routes=[MyAPIView]) as client:
+def test_cannot_access_controller(test_client_factory):
+    with create_client(routes=[MyController]) as client:
         response = client.get("/home")
 
         assert response.status_code == 403
-        assert MyAPIView.permissions == [DenyAll]
+        assert MyController.permissions == [DenyAll]
 
 
-def test_cannot_access_simple_apiview():
-    with create_client(routes=[MySimpleAPIView]) as client:
+def test_cannot_access_simple_controller():
+    with create_client(routes=[MySimpleController]) as client:
         response = client.get("/new-home")
 
         assert response.status_code == 403
-        assert MySimpleAPIView.permissions == [DenyAll, AllowAny]
+        assert MySimpleController.permissions == [DenyAll, AllowAny]
 
 
 def test_inheritance_total(test_client_factory):